Seperate JWT logic from GraphQL
