+++ graphql/spec/controllers/spree/graphql_controller_spec.rb
@@ -1,23 +1,8 @@
 require 'spec_helper'
 
-describe Spree::Graphql::GraphqlController, type: :controller do
+describe Spree::GraphqlController, type: :controller do
   # TODO: [SGQL-1] Add specs for failed token and access denied
   context 'JWT token access' do
-    context 'login' do
-      let!(:user) { create :user }
-
-      before do
-        allow(controller).to receive(:spree_current_user).and_return(user)
-      end
-
-      it do
-        post :login, params: { email: user.email, password: user.password }
-        expect(json_response['token']).to be_present
-        expect(json_response['exp']).to be_present
-        expect(json_response['login']).to be_present
-      end
-    end
-
     context 'authorized request' do
       let(:user) { create :user }
       let(:access_token) { Spree::JwtToken.create_for_user(user)[:token] }
