+++ graphql/spec/support/json_response.rb
@@ -1,0 +1,8 @@
+def json_response
+  case body = JSON.parse(response.body)
+  when Hash
+    body.with_indifferent_access
+  when Array
+    body
+  end
+end
